refactor: migrate product sync to v2
